✅ test: hero background, three.js 테스트 코드 분리
